Seperate op_utils and utils
